move validations parity check to tests directory
