--- src/core/validations/client_parity_tests/mod.rs
@@ -1,2 +1,0 @@
-#[cfg(test)]
-mod test;
